+++ test/lib/webrtc/detectsilence.js
