+++ test/lib/webrtc/fakemediastream.js
