+++ test/lib/webrtc/fakestats.js
