--- test/webrtc/unit/index.js
@@ -1,8 +1,0 @@
-'use strict';
-
-require('./spec/getstats');
-require('./spec/getusermedia');
-
-require('./spec/util');
-require('./spec/util/latch');
-require('./spec/util/sdp');
